requirements.txt
discord.py==1.7.3
discord-py-slash-command==2.0.0
asyncio==3.4.3
